requirements.txt
fastapi[all]
sqlalchemy[asyncio]
python-dotenv 
psycopg2-binary